rename data related files
